🚚 refactor: Changed the upload module directory to inside the materials module directory.
